@@ -4,27 +4,26 @@ version: 0.0.1
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   data:
     path: ../data/
-  injectable: ^1.5.3
-  drift: ^1.3.0
-  sqlite3_flutter_libs: ^0.5.8
-  path_provider: ^2.0.11
-  path: ^1.8.0
+  injectable: ^2.5.0
+  drift: ^2.28.0
+  sqlite3_flutter_libs: ^0.5.34
+  path_provider: ^2.1.5
+  path: ^1.9.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  injectable_generator: ^1.5.3
-  build_runner: ^2.2.0
-  drift_dev: ^1.3.0
-  flutter_lints: ^2.0.1
-  mockito: ^5.2.0
+  injectable_generator: ^2.7.0
+  build_runner: ^2.5.4
+  drift_dev: ^2.28.0
+  flutter_lints: ^6.0.0
+  mockito: ^5.4.6
 
 flutter:
