@@ -4,22 +4,21 @@ version: 0.0.1
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   domain:
     path: ../domain/
-  injectable: ^1.5.3
+  injectable: ^2.5.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  injectable_generator: ^1.5.3
-  build_runner: ^2.2.0
-  flutter_lints: ^2.0.1
-  mockito: ^5.2.0
+  injectable_generator: ^2.7.0
+  build_runner: ^2.5.4
+  flutter_lints: ^6.0.0
+  mockito: ^5.4.6
 
 flutter:
