@@ -1,21 +1,21 @@
 name: flutter_clean_architecture
-description: A Flutter project to demonestrate clean architecture implementation.
+description: A Flutter project to demonstrate clean architecture implementation.
 publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^1.0.5
-  equatable: ^2.0.3
-  flutter_bloc: ^8.0.1
-  get_it: ^7.2.0
-  injectable: ^1.5.3
-  url_launcher: ^6.1.5
+  cupertino_icons: ^1.0.8
+  equatable: ^2.0.7
+  flutter_bloc: ^9.1.1
+  get_it: ^8.0.3
+  injectable: ^2.5.0
+  url_launcher: ^6.3.2
   remote:
     path: ./remote/
   cache:
@@ -28,10 +28,10 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  injectable_generator: ^1.5.2
-  build_runner: ^2.2.0
-  bloc_test: ^9.0.3
-  flutter_lints: ^2.0.1
+  injectable_generator: ^2.7.0
+  build_runner: ^2.5.4
+  bloc_test: ^10.0.0
+  flutter_lints: ^6.0.0
 
 flutter:
   uses-material-design: true
