@@ -14,13 +14,11 @@ class EndpointProvider {
   Future<ArticlesResponseModel> getMostEmailedArticles() async {
     try {
       final response = await _dio.get('mostpopular/v2/emailed/7.json',
-          queryParameters: <String, String>{
-            'api-key': 'ADD-YOUR-KEY'
-          });
+          queryParameters: <String, String>{'api-key': 'ADD-YOUR-KEY'});
 
       return ArticlesResponseModel.fromJson(json.decode(response.toString()));
-    } on DioError catch (ex) {
-      developer.log(ex.message, stackTrace: ex.stackTrace);
+    } on DioException catch (ex) {
+      developer.log(ex.message ?? "", stackTrace: ex.stackTrace);
       throw Exception(ex.message);
     }
   }
@@ -28,13 +26,11 @@ class EndpointProvider {
   Future<ArticlesResponseModel> getMostViewedArticles() async {
     try {
       final response = await _dio.get('mostpopular/v2/viewed/7.json',
-          queryParameters: <String, String>{
-            'api-key': 'ADD-YOUR-KEY'
-          });
+          queryParameters: <String, String>{'api-key': 'ADD-YOUR-KEY'});
 
       return ArticlesResponseModel.fromJson(json.decode(response.toString()));
-    } on DioError catch (ex) {
-      developer.log(ex.message, stackTrace: ex.stackTrace);
+    } on DioException catch (ex) {
+      developer.log(ex.message ?? "", stackTrace: ex.stackTrace);
       throw Exception(ex.message);
     }
   }
@@ -42,13 +38,11 @@ class EndpointProvider {
   Future<ArticlesResponseModel> getMostSharedArticles() async {
     try {
       final response = await _dio.get('mostpopular/v2/shared/7.json',
-          queryParameters: <String, String>{
-            'api-key': 'ADD-YOUR-KEY'
-          });
+          queryParameters: <String, String>{'api-key': 'ADD-YOUR-KEY'});
 
       return ArticlesResponseModel.fromJson(json.decode(response.toString()));
-    } on DioError catch (ex) {
-      developer.log(ex.message, stackTrace: ex.stackTrace);
+    } on DioException catch (ex) {
+      developer.log(ex.message ?? "", stackTrace: ex.stackTrace);
       throw Exception(ex.message);
     }
   }
