@@ -2,27 +2,90 @@
 
 part of 'demo_db.dart';
 
-// **************************************************************************
-// MoorGenerator
-// **************************************************************************
+// ignore_for_file: type=lint
+class $ArticlesEntityTable extends ArticlesEntity
+    with TableInfo<$ArticlesEntityTable, Articles> {
+  @override
+  final GeneratedDatabase attachedDatabase;
+  final String? _alias;
+  $ArticlesEntityTable(this.attachedDatabase, [this._alias]);
+  static const VerificationMeta _contentMeta =
+      const VerificationMeta('content');
+  @override
+  late final GeneratedColumn<String> content = GeneratedColumn<String>(
+      'content', aliasedName, false,
+      type: DriftSqlType.string, requiredDuringInsert: true);
+  static const VerificationMeta _typeMeta = const VerificationMeta('type');
+  @override
+  late final GeneratedColumn<String> type = GeneratedColumn<String>(
+      'type', aliasedName, false,
+      type: DriftSqlType.string, requiredDuringInsert: true);
+  static const VerificationMeta _updatedMeta =
+      const VerificationMeta('updated');
+  @override
+  late final GeneratedColumn<DateTime> updated = GeneratedColumn<DateTime>(
+      'updated', aliasedName, false,
+      type: DriftSqlType.dateTime, requiredDuringInsert: true);
+  @override
+  List<GeneratedColumn> get $columns => [content, type, updated];
+  @override
+  String get aliasedName => _alias ?? actualTableName;
+  @override
+  String get actualTableName => $name;
+  static const String $name = 'articles_entity';
+  @override
+  VerificationContext validateIntegrity(Insertable<Articles> instance,
+      {bool isInserting = false}) {
+    final context = VerificationContext();
+    final data = instance.toColumns(true);
+    if (data.containsKey('content')) {
+      context.handle(_contentMeta,
+          content.isAcceptableOrUnknown(data['content']!, _contentMeta));
+    } else if (isInserting) {
+      context.missing(_contentMeta);
+    }
+    if (data.containsKey('type')) {
+      context.handle(
+          _typeMeta, type.isAcceptableOrUnknown(data['type']!, _typeMeta));
+    } else if (isInserting) {
+      context.missing(_typeMeta);
+    }
+    if (data.containsKey('updated')) {
+      context.handle(_updatedMeta,
+          updated.isAcceptableOrUnknown(data['updated']!, _updatedMeta));
+    } else if (isInserting) {
+      context.missing(_updatedMeta);
+    }
+    return context;
+  }
 
-// ignore_for_file: unnecessary_brace_in_string_interps, unnecessary_this
+  @override
+  Set<GeneratedColumn> get $primaryKey => {type};
+  @override
+  Articles map(Map<String, dynamic> data, {String? tablePrefix}) {
+    final effectivePrefix = tablePrefix != null ? '$tablePrefix.' : '';
+    return Articles(
+      content: attachedDatabase.typeMapping
+          .read(DriftSqlType.string, data['${effectivePrefix}content'])!,
+      type: attachedDatabase.typeMapping
+          .read(DriftSqlType.string, data['${effectivePrefix}type'])!,
+      updated: attachedDatabase.typeMapping
+          .read(DriftSqlType.dateTime, data['${effectivePrefix}updated'])!,
+    );
+  }
+
+  @override
+  $ArticlesEntityTable createAlias(String alias) {
+    return $ArticlesEntityTable(attachedDatabase, alias);
+  }
+}
+
 class Articles extends DataClass implements Insertable<Articles> {
   final String content;
   final String type;
   final DateTime updated;
-  Articles({required this.content, required this.type, required this.updated});
-  factory Articles.fromData(Map<String, dynamic> data, {String? prefix}) {
-    final effectivePrefix = prefix ?? '';
-    return Articles(
-      content: const StringType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}content'])!,
-      type: const StringType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}type'])!,
-      updated: const DateTimeType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}updated'])!,
-    );
-  }
+  const Articles(
+      {required this.content, required this.type, required this.updated});
   @override
   Map<String, Expression> toColumns(bool nullToAbsent) {
     final map = <String, Expression>{};
@@ -65,6 +128,14 @@ class Articles extends DataClass implements Insertable<Articles> {
         type: type ?? this.type,
         updated: updated ?? this.updated,
       );
+  Articles copyWithCompanion(ArticlesEntityCompanion data) {
+    return Articles(
+      content: data.content.present ? data.content.value : this.content,
+      type: data.type.present ? data.type.value : this.type,
+      updated: data.updated.present ? data.updated.value : this.updated,
+    );
+  }
+
   @override
   String toString() {
     return (StringBuffer('Articles(')
@@ -90,15 +161,18 @@ class ArticlesEntityCompanion extends UpdateCompanion<Articles> {
   final Value<String> content;
   final Value<String> type;
   final Value<DateTime> updated;
+  final Value<int> rowid;
   const ArticlesEntityCompanion({
     this.content = const Value.absent(),
     this.type = const Value.absent(),
     this.updated = const Value.absent(),
+    this.rowid = const Value.absent(),
   });
   ArticlesEntityCompanion.insert({
     required String content,
     required String type,
     required DateTime updated,
+    this.rowid = const Value.absent(),
   })  : content = Value(content),
         type = Value(type),
         updated = Value(updated);
@@ -106,20 +180,26 @@ class ArticlesEntityCompanion extends UpdateCompanion<Articles> {
     Expression<String>? content,
     Expression<String>? type,
     Expression<DateTime>? updated,
+    Expression<int>? rowid,
   }) {
     return RawValuesInsertable({
       if (content != null) 'content': content,
       if (type != null) 'type': type,
       if (updated != null) 'updated': updated,
+      if (rowid != null) 'rowid': rowid,
     });
   }
 
   ArticlesEntityCompanion copyWith(
-      {Value<String>? content, Value<String>? type, Value<DateTime>? updated}) {
+      {Value<String>? content,
+      Value<String>? type,
+      Value<DateTime>? updated,
+      Value<int>? rowid}) {
     return ArticlesEntityCompanion(
       content: content ?? this.content,
       type: type ?? this.type,
       updated: updated ?? this.updated,
+      rowid: rowid ?? this.rowid,
     );
   }
 
@@ -135,6 +215,9 @@ class ArticlesEntityCompanion extends UpdateCompanion<Articles> {
     if (updated.present) {
       map['updated'] = Variable<DateTime>(updated.value);
     }
+    if (rowid.present) {
+      map['rowid'] = Variable<int>(rowid.value);
+    }
     return map;
   }
 
@@ -143,84 +226,166 @@ class ArticlesEntityCompanion extends UpdateCompanion<Articles> {
     return (StringBuffer('ArticlesEntityCompanion(')
           ..write('content: $content, ')
           ..write('type: $type, ')
-          ..write('updated: $updated')
+          ..write('updated: $updated, ')
+          ..write('rowid: $rowid')
           ..write(')'))
         .toString();
   }
 }
 
-class $ArticlesEntityTable extends ArticlesEntity
-    with TableInfo<$ArticlesEntityTable, Articles> {
-  final GeneratedDatabase _db;
-  final String? _alias;
-  $ArticlesEntityTable(this._db, [this._alias]);
-  final VerificationMeta _contentMeta = const VerificationMeta('content');
-  @override
-  late final GeneratedColumn<String?> content = GeneratedColumn<String?>(
-      'content', aliasedName, false,
-      type: const StringType(), requiredDuringInsert: true);
-  final VerificationMeta _typeMeta = const VerificationMeta('type');
-  @override
-  late final GeneratedColumn<String?> type = GeneratedColumn<String?>(
-      'type', aliasedName, false,
-      type: const StringType(), requiredDuringInsert: true);
-  final VerificationMeta _updatedMeta = const VerificationMeta('updated');
-  @override
-  late final GeneratedColumn<DateTime?> updated = GeneratedColumn<DateTime?>(
-      'updated', aliasedName, false,
-      type: const IntType(), requiredDuringInsert: true);
-  @override
-  List<GeneratedColumn> get $columns => [content, type, updated];
-  @override
-  String get aliasedName => _alias ?? 'articles_entity';
-  @override
-  String get actualTableName => 'articles_entity';
-  @override
-  VerificationContext validateIntegrity(Insertable<Articles> instance,
-      {bool isInserting = false}) {
-    final context = VerificationContext();
-    final data = instance.toColumns(true);
-    if (data.containsKey('content')) {
-      context.handle(_contentMeta,
-          content.isAcceptableOrUnknown(data['content']!, _contentMeta));
-    } else if (isInserting) {
-      context.missing(_contentMeta);
-    }
-    if (data.containsKey('type')) {
-      context.handle(
-          _typeMeta, type.isAcceptableOrUnknown(data['type']!, _typeMeta));
-    } else if (isInserting) {
-      context.missing(_typeMeta);
-    }
-    if (data.containsKey('updated')) {
-      context.handle(_updatedMeta,
-          updated.isAcceptableOrUnknown(data['updated']!, _updatedMeta));
-    } else if (isInserting) {
-      context.missing(_updatedMeta);
-    }
-    return context;
-  }
-
-  @override
-  Set<GeneratedColumn> get $primaryKey => {type};
-  @override
-  Articles map(Map<String, dynamic> data, {String? tablePrefix}) {
-    return Articles.fromData(data,
-        prefix: tablePrefix != null ? '$tablePrefix.' : null);
-  }
-
-  @override
-  $ArticlesEntityTable createAlias(String alias) {
-    return $ArticlesEntityTable(_db, alias);
-  }
-}
-
 abstract class _$DemoDB extends GeneratedDatabase {
-  _$DemoDB(QueryExecutor e) : super(SqlTypeSystem.defaultInstance, e);
+  _$DemoDB(QueryExecutor e) : super(e);
+  $DemoDBManager get managers => $DemoDBManager(this);
   late final $ArticlesEntityTable articlesEntity = $ArticlesEntityTable(this);
   late final ArticlesDao articlesDao = ArticlesDao(this as DemoDB);
   @override
-  Iterable<TableInfo> get allTables => allSchemaEntities.whereType<TableInfo>();
+  Iterable<TableInfo<Table, Object?>> get allTables =>
+      allSchemaEntities.whereType<TableInfo<Table, Object?>>();
   @override
   List<DatabaseSchemaEntity> get allSchemaEntities => [articlesEntity];
+}
+
+typedef $$ArticlesEntityTableCreateCompanionBuilder = ArticlesEntityCompanion
+    Function({
+  required String content,
+  required String type,
+  required DateTime updated,
+  Value<int> rowid,
+});
+typedef $$ArticlesEntityTableUpdateCompanionBuilder = ArticlesEntityCompanion
+    Function({
+  Value<String> content,
+  Value<String> type,
+  Value<DateTime> updated,
+  Value<int> rowid,
+});
+
+class $$ArticlesEntityTableFilterComposer
+    extends Composer<_$DemoDB, $ArticlesEntityTable> {
+  $$ArticlesEntityTableFilterComposer({
+    required super.$db,
+    required super.$table,
+    super.joinBuilder,
+    super.$addJoinBuilderToRootComposer,
+    super.$removeJoinBuilderFromRootComposer,
+  });
+  ColumnFilters<String> get content => $composableBuilder(
+      column: $table.content, builder: (column) => ColumnFilters(column));
+
+  ColumnFilters<String> get type => $composableBuilder(
+      column: $table.type, builder: (column) => ColumnFilters(column));
+
+  ColumnFilters<DateTime> get updated => $composableBuilder(
+      column: $table.updated, builder: (column) => ColumnFilters(column));
+}
+
+class $$ArticlesEntityTableOrderingComposer
+    extends Composer<_$DemoDB, $ArticlesEntityTable> {
+  $$ArticlesEntityTableOrderingComposer({
+    required super.$db,
+    required super.$table,
+    super.joinBuilder,
+    super.$addJoinBuilderToRootComposer,
+    super.$removeJoinBuilderFromRootComposer,
+  });
+  ColumnOrderings<String> get content => $composableBuilder(
+      column: $table.content, builder: (column) => ColumnOrderings(column));
+
+  ColumnOrderings<String> get type => $composableBuilder(
+      column: $table.type, builder: (column) => ColumnOrderings(column));
+
+  ColumnOrderings<DateTime> get updated => $composableBuilder(
+      column: $table.updated, builder: (column) => ColumnOrderings(column));
+}
+
+class $$ArticlesEntityTableAnnotationComposer
+    extends Composer<_$DemoDB, $ArticlesEntityTable> {
+  $$ArticlesEntityTableAnnotationComposer({
+    required super.$db,
+    required super.$table,
+    super.joinBuilder,
+    super.$addJoinBuilderToRootComposer,
+    super.$removeJoinBuilderFromRootComposer,
+  });
+  GeneratedColumn<String> get content =>
+      $composableBuilder(column: $table.content, builder: (column) => column);
+
+  GeneratedColumn<String> get type =>
+      $composableBuilder(column: $table.type, builder: (column) => column);
+
+  GeneratedColumn<DateTime> get updated =>
+      $composableBuilder(column: $table.updated, builder: (column) => column);
+}
+
+class $$ArticlesEntityTableTableManager extends RootTableManager<
+    _$DemoDB,
+    $ArticlesEntityTable,
+    Articles,
+    $$ArticlesEntityTableFilterComposer,
+    $$ArticlesEntityTableOrderingComposer,
+    $$ArticlesEntityTableAnnotationComposer,
+    $$ArticlesEntityTableCreateCompanionBuilder,
+    $$ArticlesEntityTableUpdateCompanionBuilder,
+    (Articles, BaseReferences<_$DemoDB, $ArticlesEntityTable, Articles>),
+    Articles,
+    PrefetchHooks Function()> {
+  $$ArticlesEntityTableTableManager(_$DemoDB db, $ArticlesEntityTable table)
+      : super(TableManagerState(
+          db: db,
+          table: table,
+          createFilteringComposer: () =>
+              $$ArticlesEntityTableFilterComposer($db: db, $table: table),
+          createOrderingComposer: () =>
+              $$ArticlesEntityTableOrderingComposer($db: db, $table: table),
+          createComputedFieldComposer: () =>
+              $$ArticlesEntityTableAnnotationComposer($db: db, $table: table),
+          updateCompanionCallback: ({
+            Value<String> content = const Value.absent(),
+            Value<String> type = const Value.absent(),
+            Value<DateTime> updated = const Value.absent(),
+            Value<int> rowid = const Value.absent(),
+          }) =>
+              ArticlesEntityCompanion(
+            content: content,
+            type: type,
+            updated: updated,
+            rowid: rowid,
+          ),
+          createCompanionCallback: ({
+            required String content,
+            required String type,
+            required DateTime updated,
+            Value<int> rowid = const Value.absent(),
+          }) =>
+              ArticlesEntityCompanion.insert(
+            content: content,
+            type: type,
+            updated: updated,
+            rowid: rowid,
+          ),
+          withReferenceMapper: (p0) => p0
+              .map((e) => (e.readTable(table), BaseReferences(db, table, e)))
+              .toList(),
+          prefetchHooksCallback: null,
+        ));
+}
+
+typedef $$ArticlesEntityTableProcessedTableManager = ProcessedTableManager<
+    _$DemoDB,
+    $ArticlesEntityTable,
+    Articles,
+    $$ArticlesEntityTableFilterComposer,
+    $$ArticlesEntityTableOrderingComposer,
+    $$ArticlesEntityTableAnnotationComposer,
+    $$ArticlesEntityTableCreateCompanionBuilder,
+    $$ArticlesEntityTableUpdateCompanionBuilder,
+    (Articles, BaseReferences<_$DemoDB, $ArticlesEntityTable, Articles>),
+    Articles,
+    PrefetchHooks Function()>;
+
+class $DemoDBManager {
+  final _$DemoDB _db;
+  $DemoDBManager(this._db);
+  $$ArticlesEntityTableTableManager get articlesEntity =>
+      $$ArticlesEntityTableTableManager(_db, _db.articlesEntity);
 }
