@@ -1,22 +1,32 @@
+// dart format width=80
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
 // **************************************************************************
 // InjectableConfigGenerator
 // **************************************************************************
 
-import 'package:get_it/get_it.dart' as _i1;
-import 'package:injectable/injectable.dart' as _i2;
+// ignore_for_file: type=lint
+// coverage:ignore-file
 
-import '../../domain.dart' as _i4;
-import '../use_cases/get_articles_use_case.dart'
-    as _i3; // ignore_for_file: unnecessary_lambdas
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'package:domain/domain.dart' as _i494;
+import 'package:domain/src/use_cases/get_articles_use_case.dart' as _i899;
+import 'package:get_it/get_it.dart' as _i174;
+import 'package:injectable/injectable.dart' as _i526;
 
-// ignore_for_file: lines_longer_than_80_chars
-/// initializes the registration of provided dependencies inside of [GetIt]
-_i1.GetIt $initDomainGetIt(_i1.GetIt get,
-    {String? environment, _i2.EnvironmentFilter? environmentFilter}) {
-  final gh = _i2.GetItHelper(get, environment, environmentFilter);
-  gh.singleton<_i3.GetArticlesUseCase>(
-      _i3.GetArticlesUseCase(get<_i4.ArticlesRepository>()));
-  return get;
+extension GetItInjectableX on _i174.GetIt {
+// initializes the registration of main-scope dependencies inside of GetIt
+  _i174.GetIt $initDomainGetIt({
+    String? environment,
+    _i526.EnvironmentFilter? environmentFilter,
+  }) {
+    final gh = _i526.GetItHelper(
+      this,
+      environment,
+      environmentFilter,
+    );
+    gh.singleton<_i899.GetArticlesUseCase>(
+        () => _i899.GetArticlesUseCase(gh<_i494.ArticlesRepository>()));
+    return this;
+  }
 }
