@@ -1,17 +1,28 @@
+// dart format width=80
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
 // **************************************************************************
 // InjectableConfigGenerator
 // **************************************************************************
 
-import 'package:get_it/get_it.dart' as _i1;
-import 'package:injectable/injectable.dart'
-    as _i2; // ignore_for_file: unnecessary_lambdas
+// ignore_for_file: type=lint
+// coverage:ignore-file
 
-// ignore_for_file: lines_longer_than_80_chars
-/// initializes the registration of provided dependencies inside of [GetIt]
-_i1.GetIt $initGetIt(_i1.GetIt get,
-    {String? environment, _i2.EnvironmentFilter? environmentFilter}) {
-  final gh = _i2.GetItHelper(get, environment, environmentFilter);
-  return get;
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'package:get_it/get_it.dart' as _i174;
+import 'package:injectable/injectable.dart' as _i526;
+
+extension GetItInjectableX on _i174.GetIt {
+// initializes the registration of main-scope dependencies inside of GetIt
+  _i174.GetIt $initGetIt({
+    String? environment,
+    _i526.EnvironmentFilter? environmentFilter,
+  }) {
+    _i526.GetItHelper(
+      this,
+      environment,
+      environmentFilter,
+    );
+    return this;
+  }
 }
