@@ -1,37 +1,49 @@
+// dart format width=80
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
 // **************************************************************************
 // InjectableConfigGenerator
 // **************************************************************************
 
-import 'package:data/data.dart' as _i7;
-import 'package:dio/dio.dart' as _i4;
-import 'package:get_it/get_it.dart' as _i1;
-import 'package:injectable/injectable.dart' as _i2;
+// ignore_for_file: type=lint
+// coverage:ignore-file
 
-import '../api/api_service.dart' as _i9;
-import '../api/articles_service.dart' as _i6;
-import '../api/endpoint_provider.dart' as _i5;
-import '../articles_remote_repository_impl.dart' as _i8;
-import '../mapper/articles_entity_mapper.dart'
-    as _i3; // ignore_for_file: unnecessary_lambdas
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'package:data/data.dart' as _i437;
+import 'package:dio/dio.dart' as _i361;
+import 'package:get_it/get_it.dart' as _i174;
+import 'package:injectable/injectable.dart' as _i526;
+import 'package:remote/src/api/api_service.dart' as _i569;
+import 'package:remote/src/api/articles_service.dart' as _i1031;
+import 'package:remote/src/api/endpoint_provider.dart' as _i223;
+import 'package:remote/src/articles_remote_repository_impl.dart' as _i1035;
+import 'package:remote/src/mapper/articles_entity_mapper.dart' as _i799;
 
-// ignore_for_file: lines_longer_than_80_chars
-/// initializes the registration of provided dependencies inside of [GetIt]
-_i1.GetIt $initRemoteGetIt(_i1.GetIt get,
-    {String? environment, _i2.EnvironmentFilter? environmentFilter}) {
-  final gh = _i2.GetItHelper(get, environment, environmentFilter);
-  final apiService = _$ApiService();
-  gh.factory<_i3.ArticlesEntityMapper>(() => _i3.ArticlesEntityMapper());
-  gh.singleton<_i4.Dio>(apiService.init());
-  gh.singleton<_i5.EndpointProvider>(_i5.EndpointProvider(get<_i4.Dio>()));
-  gh.singleton<_i6.ArticlesService>(
-      _i6.ArticlesService(get<_i5.EndpointProvider>()));
-  gh.factory<_i7.ArticlesRemoteRepository>(() =>
-      _i8.ArticlesRemoteRepositoryImpl(
-          articlesService: get<_i6.ArticlesService>(),
-          articlesEntityMapper: get<_i3.ArticlesEntityMapper>()));
-  return get;
+extension GetItInjectableX on _i174.GetIt {
+// initializes the registration of main-scope dependencies inside of GetIt
+  _i174.GetIt $initRemoteGetIt({
+    String? environment,
+    _i526.EnvironmentFilter? environmentFilter,
+  }) {
+    final gh = _i526.GetItHelper(
+      this,
+      environment,
+      environmentFilter,
+    );
+    final apiService = _$ApiService();
+    gh.factory<_i799.ArticlesEntityMapper>(() => _i799.ArticlesEntityMapper());
+    gh.singleton<_i361.Dio>(() => apiService.init());
+    gh.singleton<_i223.EndpointProvider>(
+        () => _i223.EndpointProvider(gh<_i361.Dio>()));
+    gh.singleton<_i1031.ArticlesService>(
+        () => _i1031.ArticlesService(gh<_i223.EndpointProvider>()));
+    gh.factory<_i437.ArticlesRemoteRepository>(
+        () => _i1035.ArticlesRemoteRepositoryImpl(
+              articlesService: gh<_i1031.ArticlesService>(),
+              articlesEntityMapper: gh<_i799.ArticlesEntityMapper>(),
+            ));
+    return this;
+  }
 }
 
-class _$ApiService extends _i9.ApiService {}
+class _$ApiService extends _i569.ApiService {}
