@@ -1,7 +1,8 @@
-// Mocks generated by Mockito 5.0.17 from annotations
+// Mocks generated by Mockito 5.4.6 from annotations
 // in remote/test/src/articles_remote_repository_impl_test.dart.
 // Do not manually edit this file.
 
+// ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i6;
 
 import 'package:data/data.dart' as _i4;
@@ -12,22 +13,52 @@ import 'package:remote/src/mapper/articles_entity_mapper.dart' as _i7;
 import 'package:remote/src/models/article/article_model.dart' as _i8;
 import 'package:remote/src/models/article/articles_response_model.dart' as _i3;
 
+// ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
+// ignore_for_file: must_be_immutable
 // ignore_for_file: prefer_const_constructors
 // ignore_for_file: unnecessary_parenthesis
 // ignore_for_file: camel_case_types
+// ignore_for_file: subtype_of_sealed_class
 
-class _FakeEndpointProvider_0 extends _i1.Fake implements _i2.EndpointProvider {
+class _FakeEndpointProvider_0 extends _i1.SmartFake
+    implements _i2.EndpointProvider {
+  _FakeEndpointProvider_0(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
 }
 
-class _FakeArticlesResponseModel_1 extends _i1.Fake
-    implements _i3.ArticlesResponseModel {}
+class _FakeArticlesResponseModel_1 extends _i1.SmartFake
+    implements _i3.ArticlesResponseModel {
+  _FakeArticlesResponseModel_1(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
-class _FakeArticlesEntity_2 extends _i1.Fake implements _i4.ArticlesEntity {}
+class _FakeArticlesEntity_2 extends _i1.SmartFake
+    implements _i4.ArticlesEntity {
+  _FakeArticlesEntity_2(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
 /// A class which mocks [ArticlesService].
 ///
@@ -38,27 +69,64 @@ class MockArticlesService extends _i1.Mock implements _i5.ArticlesService {
   }
 
   @override
-  _i2.EndpointProvider get endpointProvider =>
-      (super.noSuchMethod(Invocation.getter(#endpointProvider),
-          returnValue: _FakeEndpointProvider_0()) as _i2.EndpointProvider);
+  _i2.EndpointProvider get endpointProvider => (super.noSuchMethod(
+        Invocation.getter(#endpointProvider),
+        returnValue: _FakeEndpointProvider_0(
+          this,
+          Invocation.getter(#endpointProvider),
+        ),
+      ) as _i2.EndpointProvider);
+
   @override
   _i6.Future<_i3.ArticlesResponseModel> getMostEmailedArticles() =>
-      (super.noSuchMethod(Invocation.method(#getMostEmailedArticles, []),
-              returnValue: Future<_i3.ArticlesResponseModel>.value(
-                  _FakeArticlesResponseModel_1()))
-          as _i6.Future<_i3.ArticlesResponseModel>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #getMostEmailedArticles,
+          [],
+        ),
+        returnValue: _i6.Future<_i3.ArticlesResponseModel>.value(
+            _FakeArticlesResponseModel_1(
+          this,
+          Invocation.method(
+            #getMostEmailedArticles,
+            [],
+          ),
+        )),
+      ) as _i6.Future<_i3.ArticlesResponseModel>);
+
   @override
   _i6.Future<_i3.ArticlesResponseModel> getMostSharedArticles() =>
-      (super.noSuchMethod(Invocation.method(#getMostSharedArticles, []),
-              returnValue: Future<_i3.ArticlesResponseModel>.value(
-                  _FakeArticlesResponseModel_1()))
-          as _i6.Future<_i3.ArticlesResponseModel>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #getMostSharedArticles,
+          [],
+        ),
+        returnValue: _i6.Future<_i3.ArticlesResponseModel>.value(
+            _FakeArticlesResponseModel_1(
+          this,
+          Invocation.method(
+            #getMostSharedArticles,
+            [],
+          ),
+        )),
+      ) as _i6.Future<_i3.ArticlesResponseModel>);
+
   @override
   _i6.Future<_i3.ArticlesResponseModel> getMostViewedArticles() =>
-      (super.noSuchMethod(Invocation.method(#getMostViewedArticles, []),
-              returnValue: Future<_i3.ArticlesResponseModel>.value(
-                  _FakeArticlesResponseModel_1()))
-          as _i6.Future<_i3.ArticlesResponseModel>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #getMostViewedArticles,
+          [],
+        ),
+        returnValue: _i6.Future<_i3.ArticlesResponseModel>.value(
+            _FakeArticlesResponseModel_1(
+          this,
+          Invocation.method(
+            #getMostViewedArticles,
+            [],
+          ),
+        )),
+      ) as _i6.Future<_i3.ArticlesResponseModel>);
 }
 
 /// A class which mocks [ArticlesEntityMapper].
@@ -72,6 +140,17 @@ class MockArticlesEntityMapper extends _i1.Mock
 
   @override
   _i4.ArticlesEntity mapFromModel(_i8.ArticleModel? articleModel) =>
-      (super.noSuchMethod(Invocation.method(#mapFromModel, [articleModel]),
-          returnValue: _FakeArticlesEntity_2()) as _i4.ArticlesEntity);
+      (super.noSuchMethod(
+        Invocation.method(
+          #mapFromModel,
+          [articleModel],
+        ),
+        returnValue: _FakeArticlesEntity_2(
+          this,
+          Invocation.method(
+            #mapFromModel,
+            [articleModel],
+          ),
+        ),
+      ) as _i4.ArticlesEntity);
 }
